requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
streamlit
python-dotenv
